部门预算改为double型

+++ 电信运营商云管平台工单处理子模块1.0/back2.0/src/main/java/com/example/back2/controller/DepartmentController.java
@@ -21,7 +21,7 @@ public class DepartmentController {
      * @return 获取该部门预算
      */
     @GetMapping("getDepBudget")
-    public Integer getDepBudget(Integer depNum){
+    public Double getDepBudget(Integer depNum){
         Department department = departmentService.queryById(depNum);
         return department.getDepBudget();
     }
